Chore: Improved file structure in packages/components/src
Summary: This is to be more consistent with @kiwicom/universal-components
