+++ packages/components/src/Modal/index.js
@@ -1,0 +1,3 @@
+// @flow
+
+export { default as Modal } from './Modal';
