+++ packages/components/src/PassengersInputs/index.js
@@ -1,0 +1,3 @@
+// @flow
+
+export { default as PassengersInputs } from './PassengersInputs';
