+++ packages/components/src/Select/index.js
@@ -1,0 +1,3 @@
+// @flow
+
+export { default as Select } from './Select';
