+++ packages/components/src/TouchableWithoutFeedback/__tests__/TouchableWithoutFeedback.test.js
@@ -4,7 +4,7 @@ import * as React from 'react';
 import { shallow } from 'react-native-testing-library';
 import { View } from 'react-native';
 
-import TouchableWithoutFeedback from '../TouchableWithoutFeedback';
+import { TouchableWithoutFeedback } from '..';
 
 it('renders', () => {
   expect(
