+++ packages/components/src/TouchableWithoutFeedback/index.js
@@ -1,0 +1,5 @@
+// @flow
+
+export {
+  default as TouchableWithoutFeedback,
+} from './TouchableWithoutFeedback';
