+++ packages/components/src/TripInput/__tests__/TripInput.test.js
