+++ packages/components/src/TripInput/index.js
@@ -1,0 +1,3 @@
+// @flow
+
+export { default as TripInput } from './TripInput';
