--- packages/components/src/modal/ModalTypes.js
@@ -1,12 +1,0 @@
-// @flow
-
-export type ModalProps = {|
-  +visible: boolean,
-  +onClose: () => void,
-  +children: React.Element<any>,
-|};
-
-export type ModalContentProps = {|
-  +onClose: () => void,
-  +children: React.Element<any>,
-|};
