--- packages/components/src/passengersInputs/PassengersInputsTypes.js
@@ -1,15 +1,0 @@
-// @flow
-
-export type Props = {|
-  +adults: number,
-  +infants: number,
-  +bags: number,
-  +onClosePress: () => void,
-  +onSavePress: (passengersData: State) => void,
-|};
-
-export type State = {|
-  adults: number,
-  infants: number,
-  bags: number,
-|};
